+++ src/components/navigation/Breadcrumbs/BreadcrumbMenu/index.tsx
@@ -1,5 +1,6 @@
 import { Stack } from "@layouts/Stack";
-import { BreadcrumbMenuLink } from "@navigation/BreadcrumbMenuLink";
+
+import { BreadcrumbMenuLink } from "../BreadcrumbMenuLink";
 import { StyledBreadcrumbMenu } from "./styles";
 
 export interface IBreadcrumbRoute {
